Update breaking-changes-2017.md
Hi Team, I found a bit of unnatural sentences in Japanese language, and I would like to fix it. This proposal is supposed to be addressed as a Linguistic Suggestion from my point of view. It would be appreciated if you would check this proposal. Please let me know if you have any concerns.

This request is in regard to:

- [Changes in Visual Studio 2017 extensibility](https://docs.microsoft.com/en-us/visualstudio/extensibility/breaking-changes-2017)
- [Visual Studio 2017 機能拡張の変更](https://docs.microsoft.com/ja-jp/visualstudio/extensibility/breaking-changes-2017)

Best regards,